refactor : 나머지 화면 리팩토링 (파일 추출, navController 제거)
