+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/show/ShowDetailContentScreen.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
@@ -31,26 +32,8 @@ import com.nexters.boolti.presentation.theme.Grey10
 import com.nexters.boolti.presentation.theme.Grey30
 import com.nexters.boolti.presentation.theme.marginHorizontal
 
-fun NavGraphBuilder.ShowDetailContentScreen(
-    navController: NavController,
-    modifier: Modifier = Modifier,
-) {
-    composable(
-        route = "content",
-    ) { entry ->
-        val showViewModel: ShowDetailViewModel =
-            entry.sharedViewModel(navController = navController)
-
-        ShowDetailContentScreen(
-            modifier = modifier,
-            viewModel = showViewModel,
-            onBackPressed = { navController.popBackStack() }
-        )
-    }
-}
-
 @Composable
-private fun ShowDetailContentScreen(
+fun ShowDetailContentScreen(
     onBackPressed: () -> Unit,
     modifier: Modifier = Modifier,
     viewModel: ShowDetailViewModel = hiltViewModel(),
