+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/show/ShowImagesScreen.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavType
@@ -36,28 +37,9 @@ import com.nexters.boolti.presentation.screen.sharedViewModel
 import net.engawapg.lib.zoomable.rememberZoomState
 import net.engawapg.lib.zoomable.zoomable
 
-fun NavGraphBuilder.ShowImagesScreen(
-    navController: NavController,
-) {
-    composable(
-        route = "images/{index}",
-        arguments = listOf(navArgument("index") { type = NavType.IntType }),
-    ) { entry ->
-        val showViewModel: ShowDetailViewModel =
-            entry.sharedViewModel(navController = navController)
-        val index = entry.arguments!!.getInt("index")
-
-        ShowImagesScreen(
-            index = index,
-            viewModel = showViewModel,
-            onBackPressed = { navController.popBackStack() },
-        )
-    }
-}
-
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-private fun ShowImagesScreen(
+fun ShowImagesScreen(
     index: Int,
     onBackPressed: () -> Unit,
     modifier: Modifier = Modifier,
